accessible/tests/mochitest/events/test_focus_canvas.html

<html>

<head>
  <title>Accessible focus testing in canvas subdom</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">

    async function doTests() {
      let focused = waitForEvent(EVENT_FOCUS, "button");
      selectAllTextAndFocus("button");
      await focused;
      testStates(getAccessible("button"), STATE_FOCUSED);

      focused = waitForEvent(EVENT_FOCUS, "textbox");
      synthesizeKey("KEY_Tab", { shiftKey: false });
      await focused;
      testStates(getAccessible("textbox"), STATE_FOCUSED);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  </script>
</head>

<body>
  <a target="_blank"
     title="Expose content in Canvas element"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=495912">
    Mozilla Bug 495912
  </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <canvas>
    <input id="button" type="button">
    <input id="textbox">
  </canvas>

  <div id="eventdump"></div>
</body>
</html>
